Tests: amemded the corresponding Python and JavaScript tests for TEST(BasicSolverTest, solversInfo).
